# improve getSelector performance