**What type of PR is this?**
/kind bug

**What this PR does / why we need it**:
improve performance of getSelector function
converts selectors of resources(replicaSet, replicationController and statefulSet) which belong to a namespace by AsSelectorPreValidated is expensive while cluster has massive resources. 
Is it more reasonable to get selectors by resource which match pod's owner's kind rather than all kinds of workloads?

**Which issue(s) this PR fixes**:

Fixes #

**Special notes for your reviewer**:
result of `go test -test.bench='BenchmarkTestSelectorSpreadPriority' -benchmem`
previous `getSelector` performance:
```
BenchmarkTestSelectorSpreadPriority/100nodes-4               399           3038886 ns/op          891875 B/op      22455 allocs/op
BenchmarkTestSelectorSpreadPriority/1000nodes-4              248           4480055 ns/op          941066 B/op      26055 allocs/op
```
new `getSelector` performance:
```
BenchmarkTestSelectorSpreadPriority/100nodes-4               780           1433495 ns/op          457925 B/op      11433 allocs/op
BenchmarkTestSelectorSpreadPriority/1000nodes-4              402           2896698 ns/op          507499 B/op      15033 allocs/op
```
result: about 50% improvement in cpu and memory consumption
**Does this PR introduce a user-facing change?**:

```release-note
None
```

**Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.**:

```docs

```