@gongguan 2019-12-12
Personal opinion: CalculateSpreadPriority aims to minimizing the number of pods belonging to the same owner rather than all kinds of workloads(replicaSet, replicationController and statefulSet) on a node.

It's more reasonable to get pod's selector by its owner rather than all kinds of workloads.