@gongguan 2019-12-12
If these changes make sense, I'll fix related unit tests later.