@tanjunchen 2019-12-12
/test pull-kubernetes-bazel-test