@alculquicondor 2019-12-12
Uhm... actually unit performance tests wouldn't give us any relevant information. We need an integration test