@gongguan 2019-12-13
> It looks like your change is doing something totally different from what the PR description says.
> 
> Could you support your change with performance test measurements? I think `priorities.BenchmarkTestSelectorSpreadPriority` is the only one we have now. But it only exercises services. We just need more tests.

add `Is it more reasonable to get selectors by resource which match pod's owner's kind rather than all kinds of workloads?` in pr description.