@alculquicondor 2019-12-13
> add `Is it more reasonable to get selectors by resource which match pod's owner's kind rather than all kinds of workloads?` in pr description.

Why is it a question? Could you run the performance test (and add any necessary one) to prove that we are actually improving the runtime?
I would describe what your PR is doing like: only obtain replicas/controllers/statefulsets if a pod has owner references to them.

Also, since we are already iterating through the owner references, we could instead directly get the replicas/controllers/sss from the lister, instead of iterating and matching selectors (which is what the implementation of `GetPodFoos(pod)` does). But again, we need to confirm that this is actually faster. And we can't do this with unit tests, only integration or E2E tests would give us accurate results.