@gongguan 2019-12-15
> Why is it a question? Could you run the performance test (and add any necessary one) to prove that we are actually improving the runtime?
> I would describe what your PR is doing like: only obtain replicas/controllers/statefulsets if a pod has owner references to them.
> 
> Also, since we are already iterating through the owner references, we could instead directly get the replicas/controllers/sss from the lister, instead of iterating and matching selectors (which is what the implementation of `GetPodFoos(pod)` does). But again, we need to confirm that this is actually faster. And we can't do this with unit tests, only integration or E2E tests would give us accurate results.

Pasted result in pr description which improves nearly 50% performance. Just one concern: can we ignore replicaSet and replication controller's selector when we want to get related selectors from a statefulSet pod.
Ask for scheduler maintainers' views. Gentle ping @ahg-g @Huang-Wei