@alculquicondor 2019-12-16
@gongguan as I explained, the unit test that you used is not relevant. This is for 2 reasons:

- The test is very specific: it only tests one service.
- The service lister is a fake. The production service will likely have a very different performance.

Because of that, you need to add a integration test in test/integration/scheduler_perf.

> Just one concern: can we ignore replicaSet and replication controller's selector when we want to get related selectors from a statefulSet pod?

you mean that if a pod is part of a statefulset, it wouldn't be also in a replication controller or replica set? I believe your PR does that already indirectly.