@gongguan 2019-12-16
> * The test is very specific: it only tests one service.

I have add a new commit which refine the perf test, no long one service
> * The service lister is a fake. The production service will likely have a very different performance.

I think fake lister is enough to prove the improvement