@alculquicondor 2019-12-16
> I have add a follow up commit in this pr which refine the perf test, no long one service

Please add a new test instead of modifying the existing one. That one should remain equivalent to BenchmarkTestDefaultEvenPodsSpreadPriority.

> > * The service lister is a fake. The production service will likely have a very different performance.
> 
> I think fake lister is enough to prove the improvement

I don't disagree that this is going to be faster, but testing with a fake doesn't really give us much. Adding 1000 services, rs, sss, etc. is going to show that this is faster, but is this a representative cluster setup? Since this priority is a default one, likely the existing integration performance tests will already show improvement. Could you start by running those tests and see what we gained?