@liggitt 2019-12-17
> testing with a fake doesn't really give us much. Adding 1000 services, rs, sss, etc. is going to show that this is faster, but is this a representative cluster setup? Since this priority is a default one, likely the existing integration performance tests will already show improvement. Could you start by running those tests and see what we gained?

Agree that any changes targeted at performance improvements need to demonstrate those improvements pre-merge with a benchmark.