@gongguan 2019-12-18
test with only [performance test](https://github.com/kubernetes/kubernetes/blob/master/test/integration/scheduler_perf/test-performance.sh#L46) (comment [density test](https://github.com/kubernetes/kubernetes/blob/master/test/integration/scheduler_perf/test-performance.sh#L51) because it will produce a panic)

result:
previous:
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-8          	    1000	   4734653 ns/op
after improve:
BenchmarkSchedulingNodeAffinity/5000Nodes/5000Pods-8                     1000	   3971504 ns/op