@gongguan 2019-12-19
/cc @liggitt @alculquicondor 
seems this changes most on PodAffinity and PodAntiAffinity