@liggitt 2019-12-19
using https://godoc.org/golang.org/x/tools/cmd/benchcmp:

```
ignoring BenchmarkSchedulingPodAntiAffinity/500Nodes/100Pods-8: before has 1 instances, after has 0
ignoring BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8: before has 1 instances, after has 0
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkScheduling/500Nodes/500Pods-8                               1749294       1756047       +0.39%
BenchmarkScheduling/5000Nodes/5000Pods-8                             2803298       2896989       +3.34%
BenchmarkSchedulingSecrets/500Nodes/500Pods-8                        1821580       2025943       +11.22%
BenchmarkSchedulingSecrets/5000Nodes/5000Pods-8                      3087161       3687319       +19.44%
BenchmarkSchedulingInTreePVs/500Nodes/500Pods-8                      60843803      51123414      -15.98%
BenchmarkSchedulingInTreePVs/5000Nodes/5000Pods-8                    141586471     126992554     -10.31%
BenchmarkSchedulingMigratedInTreePVs/500Nodes/500Pods-8              29270762      31226602      +6.68%
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/5000Pods-8            53741000      57019357      +6.10%
BenchmarkSchedulingCSIPVs/500Nodes/500Pods-8                         31194833      22520978      -27.81%
BenchmarkSchedulingCSIPVs/5000Nodes/5000Pods-8                       61719422      55987735      -9.29%
BenchmarkSchedulingPodAffinity/500Nodes/500Pods-8                    4225022       4045884       -4.24%
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-8                  18028920      12526054      -30.52%
BenchmarkSchedulingPreferredPodAffinity/500Nodes/500Pods-8           4411154       3711780       -15.85%
BenchmarkSchedulingPreferredPodAffinity/5000Nodes/5000Pods-8         19438820      12313732      -36.65%
BenchmarkSchedulingPreferredPodAntiAffinity/500Nodes/500Pods-8       4616621       3402809       -26.29%
BenchmarkSchedulingPreferredPodAntiAffinity/5000Nodes/5000Pods-8     19045475      12629200      -33.69%
BenchmarkSchedulingNodeAffinity/500Nodes/500Pods-8                   2132278       2249140       +5.48%
BenchmarkSchedulingNodeAffinity/5000Nodes/5000Pods-8                 4338092       4095333       -5.60%
```