@gongguan 2019-12-19
```
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkScheduling/500Nodes/500Pods-8                               1856086       1679622       -9.51%
BenchmarkScheduling/5000Nodes/5000Pods-8                             2751876       2718412       -1.22%
BenchmarkSchedulingPodAntiAffinity/500Nodes/100Pods-8                3540758       3500571       -1.13%
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8              12534982      12724315      +1.51%
BenchmarkSchedulingSecrets/500Nodes/500Pods-8                        1825087       1840018       +0.82%
BenchmarkSchedulingSecrets/5000Nodes/5000Pods-8                      3176751       3147612       -0.92%
BenchmarkSchedulingInTreePVs/500Nodes/500Pods-8                      50360984      48854130      -2.99%
BenchmarkSchedulingInTreePVs/5000Nodes/5000Pods-8                    109833528     99443148      -9.46%
BenchmarkSchedulingMigratedInTreePVs/500Nodes/500Pods-8              23279662      13150524      -43.51%
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/5000Pods-8            51108166      51739641      +1.24%
BenchmarkSchedulingCSIPVs/500Nodes/500Pods-8                         24120356      13802306      -42.78%
BenchmarkSchedulingCSIPVs/5000Nodes/5000Pods-8                       57304858      51520157      -10.09%
BenchmarkSchedulingPodAffinity/500Nodes/500Pods-8                    3969510       4086248       +2.94%
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-8                  17994845      17855892      -0.77%
BenchmarkSchedulingPreferredPodAffinity/500Nodes/500Pods-8           4265245       4480969       +5.06%
BenchmarkSchedulingPreferredPodAffinity/5000Nodes/5000Pods-8         19417237      19000567      -2.15%
BenchmarkSchedulingPreferredPodAntiAffinity/500Nodes/500Pods-8       4582959       4479268       -2.26%
BenchmarkSchedulingPreferredPodAntiAffinity/5000Nodes/5000Pods-8     19340158      18599504      -3.83%
BenchmarkSchedulingNodeAffinity/500Nodes/500Pods-8                   2161435       2101483       -2.77%
BenchmarkSchedulingNodeAffinity/5000Nodes/5000Pods-8                 4410415       4304930       -2.39%

benchmark                                                            old allocs     new allocs     delta
BenchmarkScheduling/500Nodes/500Pods-8                               8402           8352           -0.60%
BenchmarkScheduling/5000Nodes/5000Pods-8                             14315          14081          -1.63%
BenchmarkSchedulingPodAntiAffinity/500Nodes/100Pods-8                15089          15312          +1.48%
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-8              44635          44714          +0.18%
BenchmarkSchedulingSecrets/500Nodes/500Pods-8                        9688           10092          +4.17%
BenchmarkSchedulingSecrets/5000Nodes/5000Pods-8                      17735          17428          -1.73%
BenchmarkSchedulingInTreePVs/500Nodes/500Pods-8                      36484          36400          -0.23%
BenchmarkSchedulingInTreePVs/5000Nodes/5000Pods-8                    68334          68646          +0.46%
BenchmarkSchedulingMigratedInTreePVs/500Nodes/500Pods-8              36109          36013          -0.27%
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/5000Pods-8            61986          62308          +0.52%
BenchmarkSchedulingCSIPVs/500Nodes/500Pods-8                         24672          25019          +1.41%
BenchmarkSchedulingCSIPVs/5000Nodes/5000Pods-8                       44220          44174          -0.10%
BenchmarkSchedulingPodAffinity/500Nodes/500Pods-8                    20622          20507          -0.56%
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-8                  85221          85551          +0.39%
BenchmarkSchedulingPreferredPodAffinity/500Nodes/500Pods-8           20665          20720          +0.27%
BenchmarkSchedulingPreferredPodAffinity/5000Nodes/5000Pods-8         85077          85010          -0.08%
BenchmarkSchedulingPreferredPodAntiAffinity/500Nodes/500Pods-8       21911          21901          -0.05%
BenchmarkSchedulingPreferredPodAntiAffinity/5000Nodes/5000Pods-8     90770          90748          -0.02%
BenchmarkSchedulingNodeAffinity/500Nodes/500Pods-8                   11134          11199          +0.58%
BenchmarkSchedulingNodeAffinity/5000Nodes/5000Pods-8                 19880          19917          +0.19%
```

I'm afraid changes in getSelector can't be measured by these benchmarks, because neither of them init a service, replicaset, replicationcontroller or statefulset.